requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.115.12
h11==0.16.0
idna==3.10
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.0
pydantic==2.11.4
pydantic_core==2.33.2
pytz==2025.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
openai==1.79.0
opencv-python-headless==4.11.0.86
opentelemetry-api==1.34.1
opentelemetry-distro==0.55b1
opentelemetry-exporter-otlp==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-instrumentation==0.55b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-semantic-conventions-ai==0.4.9